weibit/requirements.txt
Django==1.10.8
djangorestframework==3.7.3
